+++ docs/release_notes.rst
@@ -11,6 +11,7 @@ Release Notes
 **Of interest to the User**:
 
 - MR #XYZ: one-liner description
+- MR #1169: I(Qx, Qy) is saved in NXCANSAS format for biosans, gpsans, and eqsans workflows
 - MR 1168: Document scaling the detector panels in the user's guide
 - MR 1162: When reducing `gpsans` data with `direct_beam` scaling, the `direct_beam_scaling` parameter is now logged during
   the reduction process and stored in the output Nexus file at `reduction_information/special_parameters/direct_beam_scaling/value`.
@@ -20,6 +21,7 @@ Release Notes
 **Of interest to the Developer:**
 
 - MR #XYZ: one-liner description
+- MR #1169: creates src/drtsans/save_cansas.py to define CANSAS format file handling methods
 - MR #1171: update environment.yml to replace build in favor of python-build
 - MR #1165: update Mantid version to 6.11
 - MR #1158: added options in the JSON schemae to rescale whole detector panels